refactor: move is_dict_corrupt to base module
